requirements.txt
python-dateutil~=2.9.0.post0
PyYAML~=6.0.2